Fix introduction of autoloading bug
The previous commit introduced a call to `defined?` to check if a class
is loaded. However, that method does not invoke autoloading, and this
gem is generally loaded before Rails app files, so that method was
effectively always returning false. This replaces `defined?` with a
rescue statement in case ApplicationJob cannot be autoloaded.