Fix loss of region for non-default geocode results
Not all Geocoder::Result objects respond to `region`, but all respond
to `state`. This adds a mapping to the GeocodeJob to ensure
LoginActivity attributes are set from Geocoder::Result values that are
certain to exist. It fixes the bug where the region is not saved when
Geocoder's `ip_lookup` service is not set to the default `ipinfo_io`.