Store latitude/longitude coordinates (#20)
* Store latitude/longitude coordinates

Geocoder generally returns coordinates for IP address requests. This
modifies the geocoding job to store those values. Also updates the
migration template and adds upgrade instructions to the documentation.

* Refactor GeocodeJob to simplify

Use respond_to? instead of checking if columns exists. This reduces
both the class complexity and line count.

* Fix loss of region for non-default geocode results

Not all Geocoder::Result objects respond to `region`, but all respond
to `state`. This adds a mapping to the GeocodeJob to ensure
LoginActivity attributes are set from Geocoder::Result values that are
certain to exist. It fixes the bug where the region is not saved when
Geocoder's `ip_lookup` service is not set to the default `ipinfo_io`.

* Refactor to minimize changeset

Instead of using a mapping, revert to setting attributes individually
and using :try to set attributes that may not exist.